more problems solved

+++ Practice_problem03/reverse_series.java
@@ -7,7 +7,7 @@ public class reverse_series {
         System.out.println("Enter size of array : ");
         int size_n=sc.nextInt();
         double a[]=new double[size_n];
-        for(int i =0;i<size_n;i++) a[i]=sc.nextInt();
+        for(int i=0;i<size_n;i++) a[i]=sc.nextDouble();
         System.out.println("Reverse order of the array :");
         for(int i=size_n-1;i>=0;i--) System.out.print(a[i]+" ");
         sc.close();
